Refactor __init__.py to satisfy RUF067 lint rule
Move BasicConfig class, basicConfig function, and helper functions
to _basic_config.py module. Move setup_rust_logging and test hooks
to _rust_compat.py module. The __init__.py now only contains
docstrings and re-exports as required by the RUF067 rule.
